pyproject.toml
[project]
name = "ai-agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
